Refactor front end for ProjectsScreen (separate component)
